requirements.txt
# Core Dependencies
fastapi>=0.104.1
cryptography>=43.0.0
pydantic-settings>=2.0.0
uvicorn>=0.24.0
streamlit>=1.28.2
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database & Async
sqlalchemy>=2.0.23
asyncpg>=0.29.0
aiomysql>=0.2.0
pymysql>=1.1.0
asyncio-pool>=0.6.0

# LLM & AI
openai>=1.3.0
langchain>=0.0.340
langchain-community>=0.0.1
langchain-core>=0.1.0
sentence-transformers>=2.2.2

# Vector Database
chromadb>=0.4.18
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.1.4
numpy>=1.24.3
networkx>=3.2.1
plotly>=5.17.0
graphviz>=0.20.1

# Utilities
typer>=0.9.0
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.1
psutil>=5.9.6

# Development
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
pre-commit>=3.6.0

# Optional: For advanced features
neo4j>=5.15.0
redis>=5.0.1
duckdb>=0.9.2